+++ ui/main/src/app/modules/core/application-loading/application-loading.component.html
@@ -19,8 +19,6 @@
 <of-activityarea-choice-after-login #activityAreaChoiceAfterLogin>
 </of-activityarea-choice-after-login>
 
-<of-sound-activation *ngIf="applicationLoaded"> </of-sound-activation>
-
 <div id="opfab-disconnected-message" class="opfab-disconnected-message" *ngIf="applicationLoader.isDisconnected">
     <div style="text-align: center;font-size:16px;font-weight: bold;margin-top:10px" translate>
         global.disconnected
