--- ui/main/src/app/modules/core/application-loading/sound-activation/sound-activation.component.html
@@ -1,16 +1,0 @@
-<!-- Copyright (c) 2018-2022, RTE (http://www.rte-france.com)              -->
-<!-- See AUTHORS.txt                                                       -->
-<!-- This Source Code Form is subject to the terms of the Mozilla Public   -->
-<!-- License, v. 2.0. If a copy of the MPL was not distributed with this   -->
-<!-- file, You can obtain one at http://mozilla.org/MPL/2.0/.              -->
-<!-- SPDX-License-Identifier: MPL-2.0                                      -->
-<!-- This file is part of the OperatorFabric project.                      -->
-
-<ng-template #noSound let-modal>
-    <div class="modal-body text-center">
-        <p translate> global.activateSoundText </p>
-        <button id="opfab-actviate-sound" type="button" class="opfab-btn" (click)="closeModal()" translate>
-            global.activateSoundButton</button>
-
-    </div>
-</ng-template>
